Language for graphics now independent from the GUI language.
Now allows for consistent different language of outputs accross
the range of graphics produced: all the language dependent graphical
elements are now controled by the option "P.lang" (not "P.GUIlang"
any more). It is for example possible to have the interface in French
and the graphics fully translated in English (for international
publication for instance).

Text outputs, on the other hand, have been left in the GUI language.